# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "St Paul's choir attendance API",
    "version": "1.0.0",
    "description": "REST API for choir events, attendance, members, feedback, and FAQs. Local apps authenticate with httpOnly cookies. Send header X-Auth-Client: bearer on sign-in to receive token and refreshToken in JSON, then use Authorize in this UI.",
    "contact": {
      "name": "Rigin Oommen",
      "email": "riginoommen@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "The server hosting this document"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from a response that included X-Auth-Client: bearer."
      },
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "httpOnly access cookie set by login, refresh, and change-password when X-Auth-Client is omitted."
      }
    },
    "parameters": {
...
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }
      },
      "Search": {
        "name": "search",
        "in": "query",
        "schema": { "type": "string" }
      },
      "From": {
        "name": "from",
        "in": "query",
        "schema": { "type": "string", "format": "date" },
        "description": "YYYY-MM-DD. Ignored when eventId is set."
      },
      "To": {
        "name": "to",
        "in": "query",
        "schema": { "type": "string", "format": "date" },
        "description": "YYYY-MM-DD. Ignored when eventId is set."
      },
      "EventType": {
        "name": "type",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/EventType" }
      },
      "LiturgicalColorQuery": {
        "name": "liturgicalColor",
        "in": "query",
        "schema": { "type": "string" },
        "description": "white, green, purple, red, black, or __none__ for events with no color."
      },
      "Year": {
        "name": "year",
        "in": "query",
        "schema": { "type": "integer" }
      },
      "EventId": {
        "name": "eventId",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/ObjectId" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request body or query is not valid.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Sign-in is missing or the session expired.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "The account cannot perform this action yet, or the role is wrong.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Conflict": {
        "description": "Username or email is already in use.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "ObjectId": {
        "type": "string",
        "pattern": "^[a-fA-F0-9]{24}$",
        "example": "64f1a2b3c4d5e6f7a8b9c0d1"
      },
      "VoicePart": {
        "type": "string",
        "enum": ["soprano", "alto", "tenor", "bass", "other"]
      },
      "ChoirVoicePart": {
        "type": "string",
        "description": "Voice part accepted when creating an account. other remains on older records.",
        "enum": ["soprano", "alto", "tenor", "bass"]
      },
      "Role": {
        "type": "string",
        "enum": ["member", "admin"]
      },
      "ApprovalStatus": {
        "type": "string",
        "enum": ["pending", "approved", "rejected"]
      },
      "EventType": {
        "type": "string",
        "enum": ["practice", "service", "concert", "other"]
      },
      "LiturgicalColor": {
        "type": "string",
        "enum": ["white", "green", "purple", "red", "black", ""]
      },
...
      "HistoryStatusFilter": {
        "type": "string",
        "enum": ["present", "absent", "late", "excused", "upcoming", "unmarked"]
      },
...
      "FaqAudience": {
        "type": "string",
        "enum": ["member", "admin", "both"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/ObjectId" },
          "name": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" },
          "onRoster": {
            "type": "boolean",
            "description": "Always true for a member. True for an admin who still sings."
          },
          "voicePart": { "$ref": "#/components/schemas/VoicePart" },
          "active": { "type": "boolean" },
          "approvalStatus": { "$ref": "#/components/schemas/ApprovalStatus" },
          "mustChangePassword": { "type": "boolean" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "rangeStart": { "type": "integer" },
          "rangeEnd": { "type": "integer" },
          "hasPrevious": { "type": "boolean" },
          "hasNext": { "type": "boolean" }
        }
      },
      "AttendanceSummary": {
        "type": "object",
        "properties": {
          "present": { "type": "integer" },
          "absent": { "type": "integer" },
          "late": { "type": "integer" },
          "excused": { "type": "integer" },
          "total": { "type": "integer" },
          "rate": {
            "type": "integer",
            "description": "round((present + late) / (present + absent + late) * 100). Excused is excluded."
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/ObjectId" },
          "title": { "type": "string" },
          "date": { "type": "string", "format": "date-time" },
          "type": { "$ref": "#/components/schemas/EventType" },
          "notes": { "type": "string" },
          "liturgicalColor": { "$ref": "#/components/schemas/LiturgicalColor" }
        }
      },
      "EventWrite": {
        "type": "object",
        "required": ["title", "date"],
        "properties": {
          "title": { "type": "string" },
          "date": { "type": "string", "format": "date-time" },
          "type": { "$ref": "#/components/schemas/EventType" },
          "notes": { "type": "string" },
          "liturgicalColor": { "$ref": "#/components/schemas/LiturgicalColor" }
        }
      },
      "Account": {
        "allOf": [
          { "$ref": "#/components/schemas/User" },
          {
            "type": "object",
            "properties": {
              "createdAt": { "type": "string", "format": "date-time" },
              "voiceRange": { "type": "string" },
              "choirPathway": { "type": "string" },
              "summary": { "$ref": "#/components/schemas/AttendanceSummary" }
            }
          }
        ]
      },
...
      "Faq": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/ObjectId" },
          "question": { "type": "string", "maxLength": 300 },
          "answer": { "type": "string", "maxLength": 5000 },
          "audience": { "$ref": "#/components/schemas/FaqAudience" },
          "sortOrder": { "type": "number" },
          "published": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events",
        "operationId": "listEvents",
        "parameters": [
          { "$ref": "#/components/parameters/Search" },
          { "$ref": "#/components/parameters/EventType" },
          { "$ref": "#/components/parameters/LiturgicalColorQuery" },
          { "$ref": "#/components/parameters/Year" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "memberId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ObjectId" },
            "description": "Admin only. Limit events by this singer's attendance."
          },
          {
            "name": "attendanceStatus",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/HistoryStatusFilter" },
            "description": "Requires memberId."
          }
        ],
        "responses": {
          "200": {
            "description": "One page of events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Event" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "totalUnfiltered": { "type": "integer" },
                        "memberId": { "type": "string" },
                        "attendanceStatus": { "type": "string", "nullable": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create an event",
        "operationId": "createEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EventWrite" } }
          }
        },
        "responses": {
          "201": {
            "description": "Event created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "event": { "$ref": "#/components/schemas/Event" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/attendance": {
      "get": {
        "tags": ["Attendance"],
        "summary": "List attendance records",
        "description": "Admin. Optional eventId and userId filters. Not paginated.",
        "operationId": "listAttendance",
        "parameters": [
          { "$ref": "#/components/parameters/EventId" },
          {
            "name": "userId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ObjectId" }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "records": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/members": {
      "get": {
        "tags": ["Members"],
        "summary": "Approval queues and admins",
        "description": "Admin. Available even when the admin still has mustChangePassword set.",
        "operationId": "listMemberQueues",
        "responses": {
          "200": {
            "description": "Pending, inactive, and declined members, plus every admin.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pending": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } },
                    "inactive": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } },
                    "declined": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } },
                    "admins": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Members"],
        "summary": "Create an account",
        "description": "Admin. New members are approved. role defaults to member.",
        "operationId": "createMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "email", "password"],
                "properties": {
                  "name": { "type": "string", "minLength": 2 },
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8, "format": "password" },
                  "voicePart": { "$ref": "#/components/schemas/ChoirVoicePart" },
                  "role": { "$ref": "#/components/schemas/Role" },
                  "onRoster": {
                    "type": "boolean",
                    "description": "For an admin, false means they do not sing."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "member": { "$ref": "#/components/schemas/User" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/faqs": {
      "get": {
        "tags": ["FAQs"],
        "summary": "List FAQs",
        "description": "Singers see published member and both. Admins see published admin and both. manage=true returns every FAQ, including drafts.",
        "operationId": "listFaqs",
        "parameters": [
          {
            "name": "manage",
            "in": "query",
            "schema": { "type": "string", "enum": ["true"] },
            "description": "Admin only. Include unpublished FAQs."
          },
          {
            "name": "audience",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/FaqAudience" },
            "description": "Admin filter on the published list."
          }
        ],
        "responses": {
          "200": {
            "description": "FAQs in sort order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "faqs": { "type": "array", "items": { "$ref": "#/components/schemas/Faq" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["FAQs"],
        "summary": "Create a FAQ",
        "operationId": "createFaq",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question", "answer", "audience"],
                "properties": {
                  "question": { "type": "string", "maxLength": 300 },
                  "answer": { "type": "string", "maxLength": 5000 },
                  "audience": { "$ref": "#/components/schemas/FaqAudience" },
                  "sortOrder": { "type": "number" },
                  "published": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "FAQ created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "faq": { "$ref": "#/components/schemas/Faq" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] },
    { "accessCookie": [] }
  ]
}